Start working on gif encoder
