+++ src/gif/mod.rs
@@ -9,3 +9,22 @@
 pub use self::decoder::GIFDecoder;
 
 mod decoder;
+mod encoder;
+
+
+#[derive(FromPrimitive)]
+/// Known block types
+enum Block {
+    Image = 0x2C,
+    Extension = 0x21,
+    Trailer = 0x3B
+}
+
+#[derive(FromPrimitive)]
+/// Known GIF extensions
+enum Extension {
+    Text = 0x01,
+    Control = 0xF9,
+    Comment = 0xFE,
+    Application = 0xFF
+}
